+++ src/main/java/nl/ramsolutions/sw/magik/jacoco/sw5lib/Sw5LibAnalyzer.java
@@ -6,9 +6,11 @@ import org.objectweb.asm.tree.MethodNode;
 
 import javax.annotation.CheckForNull;
 
+import java.nio.file.Path;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -32,6 +34,10 @@ public final class Sw5LibAnalyzer {
      */
     public Sw5LibAnalyzer(final Sw5LibReader libReader) {
         this.libReader = libReader;
+    }
+
+    public List<Path> getProductPaths() {
+        return this.libReader.getProductPaths();
     }
 
     /**