+++ src/test/java/nl/ramsolutions/sw/magik/jacoco/helpers/SmallworldProductsSourceFileLocatorTest.java
@@ -1,4 +1,4 @@
-package nl.ramsolutions.sw.magik.jacoco.generators;
+package nl.ramsolutions.sw.magik.jacoco.helpers;
 
 import nl.ramsolutions.sw.magik.jacoco.TestData;
 import org.junit.jupiter.api.Test;
@@ -8,14 +8,15 @@ import java.io.Reader;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class MagikDirectorySourceFileLocatorTest {
+class SmallworldProductsSourceFileLocatorTest {
 
     private static final String PACKAGE_NAME = "magik/fixture_product/fixture_module";
     private static final String FILENAME = "modules/fixture_module/source/char16_vector.magik";
 
     @Test
     void testGetSourceFile() throws IOException {
-        final MagikProductSourceFileLocator locator = new MagikProductSourceFileLocator(TestData.PRODUCT_PATH);
+        final SmallworldProducts smallworldProducts = new SmallworldProducts(TestData.PRODUCT_PATHS);
+        final SmallworldProductsSourceFileLocator locator = new SmallworldProductsSourceFileLocator(smallworldProducts);
         final Reader reader = locator.getSourceFile(PACKAGE_NAME, FILENAME);
         assertThat(reader).isNotNull();
     }