+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes
 
 2.1.1 (unreleased)
 
+- Fix resolving magik file paths in sub-products.
+
 2.1.0 (2023-09-25)
 
 - Rename `--xml`-parameter to `--jacoco-xml`-parameter, and rename `XmlReportGenerator` to `JacocoXmlReportGenerator`.